feat: Testing and packaging
We'd like to support e.g. multiple python versions, and managing this is
really hard without some virtualenv and package manager.

For the virtualenv management, `tox` is used. It is a
platform-independent Python-specific tool for testing purposes.

For the packaging, `poetry` is used. `setuptools` has also been
considered, but since PEP 518, it is getting deprecated slowly. It is
sensible to use some tool which complies with PEP 518, i.e. the
specification for the `pyproject.toml` file.

A `README.md` has been created, since a development guide needs to be
established.

A simple test using `pytest` has been added, because `pytest` would fail
if no test cases were found.

A `.gitignore` file has been added to exclude pycache and tox/poetry
build files.