feat(piquasso): Interface suggestion
An initial implementation of the Piquasso programming interface has been
established.

The heart of the implementation is the `Program` class, which stores the
backend and the instuctions, which are registered to the current quantum
program.

The vectors and matrices are represented using `numpy`, therefore it is
added as a development dependency.

In `piquasso.operator` sits the `BaseOperator`, which is the base of all
the operators in this framework, including states and gates. E.g. in
`piquasso.state`, an initial quantum state implementation (`State`)
uses `BaseOperator`.

There is an initial `FockBackend` with a beamsplitter implementation,
and an empty `GaussianBackend` with no implementation whatsoever.

Tests are written (under `tests`) demonstrating this initial programming
interface.