dev: Added `flake8` linter tool
IMHO linting is important to reduce errors and improve the quality of
our code. Moreover, it is not that terribly hard to install, and
integrates well with most IDEs.

Therefore I modified the development and testing environment slightly:
- I issued `poetry add flake8`;
- fixed the current problems `flake8` showed;
- added `flake8` to `tox.ini`.

Moreover, the `tests/` directory at the `pytest` invocation is removed,
because it is redundant.

To use `flake8` CLI, just run `poetry run flake8` in the project root.