feat(gaussian.state): GaussianState
**Scope**

A minimal implementation of the gaussian state representations have been
implemented in `piquasso.gaussian.state.GaussianState`.

Note, that it **doesn't** include all transformations possible, but
provides the minimal interface for implementing measurements.

**Details**

`xp_mean` and `xp_corr` are the mean and correlation matrix in the
xp-ordering, i.e. the ordering in which the `x` and `p` operators are
separated. These are implemented as properties.

The xp-ordered mean and correlation matrix is not the final one: this
has to be transformed with `quad_transformation`, which helps
calculating the mean and correlation matrix (or covariance matrix) in
the desired representation.

References: #20