feat(gaussian): Displacement gate
The displacement gate is implemented for the `GaussianBackend` in
`GaussianBackend.displacement`.

The `D` gate has also been added to `piquasso.gates`, and the
before-mentioned `GaussianBackend.displacement` method has been
specified in it. This is done to enable the `Program` to consider the
`GaussianBackend.displacement` method as an implementation of `D` on
`GaussianBackend`.

Since using the package would require a lot of boilerplate imports as
the package grows, the new `D` gate along with several already existing
names are imported in `__init__.py` files. `__all__` is needed to
specify public names for wildcard imports.

Unit tests were written in `piquasso/gaussian/tests/test_backend.py` and
functional tests were written in `tests/functional/test_gaussian.py`.
The unit tests solely test the `GaussianBackend.displacement` method,
while the functional one tests the gate via the public interface of piquasso.

Issue: #27