feat(gaussian.state): Moving to Cython
Cython is installed with `poetry add cython`, and added as a build
dependency in `pyproject.toml`

For building a Cython extension, `setuptools` is used as a build
dependency, which is also added to `pyproject.toml`.

Also for the Cython extension, a separate `build.py` needs to be
created, where the `Extension` is created using `setuptools`, see [the
Cython
documentation](https://cython.readthedocs.io/en/latest/src/userguide/source_files_and_compilation.html).
The `build.py` is specified in `pyproject.toml`, so that the `poetry
build` command builds the Cython extension.

`piquasso/gaussian/state.py` got renamed to `state.pyx` to indicate that
it is a Cython extension. Also, the `distutils: language = c++` line got
added to the beginning of the file to indicate that it should compile to
a C++ file. `state.pyx` is also specified in `build.py`.

`.gitignore` is extended with new binaries and build files.

**Guide**

- Exit from the virtualenv.
- Run `poetry build`. This creates a pip wheel (`.whl`) file and a
  tarball `.tar.gz`.
- Install the wheel with `pip install dist/<package-name>.whl`, where
  the `<package-name>` is the name of the wheel.
- Go to a separate path, which does not contain any directory
  named `piquasso`. This is done to avoid importing the `piquasso`
  module from your working directory.
- Open up a `python3` shell, and issue
```
>>> from piquasso.gaussian.state import GaussianState
```
  which should work without errors.

References: #23