fix(apply_to_C_and_G): conjugation and indexing
**Problem**

- A conjugation of the `T` matrix has been missing at `apply_to_C_and_G`
  in the case of applying the matrix to the remaining modes. This is
  because `apply_to_C_and_G` expects a `T` matrix which transforms e.g.
  the expectation value of the annihilation operators. A more detailed
  description of the transformation of vectors of ladder operators can
  be found [in this LaTeX document](https://www.overleaf.com/read/xrkncztkffzv).

- As it turned out, `apply_to_C_and_G` only worked for certain cases,
  e.g. 2 mode gates with 3 mode states, but haven't been noticed without
  proper unit testing.

**Solution**

The computation has been extended with the conjugation. This didn't
break any of the previous tests, which was expected, but this statement
alone cries for other tests.

The computation has also been generalized for any `(l, l)` shaped `T`
matrix, where `l` is a positive integer.

Also, the edge case when there are no remaining modes (i.e. `T` is a
d×d matrix, where `d` is the number of modes) need to be also
handled, since one couldn't index a numpy array with an empty index
array.

The `apply_to_C_and_G` method got too large and complex, so another
method named `_apply_to_other_modes` got extracted from it, which deals
with the application of the transformation to the other indices of `C`
and `G` which are not handled in the caller method `apply_to_C_and_G`.

`rows` and `columns` got renamed to `transformed_rows` and
`transformed_columns`, and `all_other_*` got renamed to `other_*` for
brevity.

`TestGaussianState` has been separated into
`TestGaussianStateRepresentations` and `TestGaussianStateOperations`,
since the dimension of the two states used for the two test classes
differ for practical reasons. `TestGaussianStateRepresentations` uses a
state with `d == 2`, which render the covariance matrix a 4-by-4 matrix,
and is conveniently written in a fixture.
Using `d == 3` would make the covariance matrix a `6-by-6` matrix, which
is much more inconvenient to deal with at testing. However, for
`TestGaussianStateOperations` using `d == 3` is more sensible,
especially if we want to test non-trivial cases of `apply_to_C_and_G`.

Several test cases have been written in `TestGaussianStateOperations`
testing `apply_to_C_and_G`
- for application to 1 mode only (`T.shape` is `(1, 1)`);
- for application to 2 modes;
- for application to all the modes (3);
- for application to 3 modes, where the modes are `(0, 1, 3)`, i.e. they
  are disjoint. This test alone uses a state with `d == 5`, hence it is
  a `staticmethod`.

The latter is important, since it verifies that `apply_to_C_and_G` can
run also with disjoint modes with multiple external modes (~
`all_other_modes`).

For the latter test case, it would be cumbersome to hardcode matrices,
therefore the following fixtures were created:
- `generate_symmetric_matrix` for usage in
  `generate_complex_symmetric_matrix`,
- `generate_complex_symmetric_matrix` for generating a random `G`
  matrix,
- `generate_unitary_matrix` for usage in `generate_hermitian_matrix` and
  for generating a random transformation in
  `test_apply_to_C_and_G_for_5_modes`,
- `generate_hermitian_matrix` for generating a random `C` matrix.