ci(.gitlab-ci.yml): `buildpackage` stage
I added the `buildpackage` stage back to the pipeline in order to build
`pypi` packages. The built-in GitLab-CI credentials are used for
authenticating into the package registry.

This pipeline can only be run on the Web UI (with `only: - web`), under
`CI/CD->Pipelines->Run Pipeline` (green button). On it, select the
desired branch. Normally, one would select the `master` branch or
a git tag named after the new version.

Be aware, that one has to bump (increment) the version of the project in
order to be able to build a package: same-version packages collide on
the registry, and it is not permitted (rightfully so!) to overwrite
existing packages.

The version `0.1.0` is bumped to `0.1.1` to test this stage.

Also, `README.md` got updated according to the changes.

Finally, `test_version` has been deleted, since it is not actually
needed, it served as a dummy test at the start of the project. Since
there are no other tests in the file, `test_piquasso.py` got deleted.