refact(GaussianState): `apply` method
**Problem**

Applying a transformation to the `C` and `G` matrices are non-trivial,
therefore this calculation got its own method before. However,
transforming the `mean` is not included.

This is odd, since there should be no scenario when the `C` and `G`
matrices are transformed, but the `mean` is not.

**Solution**

A common `apply` method has been created, which transforms the state by
- transforming `mean` with a matrix multiplication and advanced
  indexing;
- calling `apply_to_C_and_G`, which deals with transforming the
  correlation matrices `C` and `G`.

The `GaussianBackend.phaseshift` method has been refactored to use this
`apply` method.

Docstring has been written to `apply`, and the rest of the docstrings
got refactored in `GaussianState`, since it was hard to reference the
instance attributes from instance methods. It is solved by the
`Attributes:` docstring keyword (Google style).

The `test_apply_to_C_and_G_for_5_modes` has been renamed to
`test_apply`, extended with the `mean` calculation through `apply`, and
moved to its own function, since it is quite independent from
`TestGaussianStateOperations`.