feat(GaussianState): Covariance matrices
The `xp_cov` and `cov` methods got implemented, which are the xp-ordered
and quadrature-ordered covariance matrices, respectively.

The tests in `test_state.py` got updated with these properties.

Copying the docstring, a typo have been found in `xp_corr` regarding
the dimensionality of the correlation matrix, and it is fixed here.

Also found some errors in the docstring regarding the density matrix,
which are also fixed.

References: #16