refact(sigma): Rename to `corr`
**Problem**

The naming `sigma` of the quadrature_ordered correlation matrix made a
lot of confusion, it is a bad naming.

**Solution**

`GaussianState.sigma` got renamed to `corr` to adhere to its meaning.

Also, the docstring has been updated accordingly.

Also, several typos have been found talking about covariance, when it
should be talking about correlation matrices. They are quite
interchangeable knowing the first moment of the distribution, but not
the same.

Let $` M `$ be the correlation matrix, $` \Sigma `$ the covariance, and
$` \mu `$ the mean. Then the following relation holds:

```math
\Sigma_{ij} = M_{ij} - 2 \mu_i \mu_j
```

References: #16