fix(xp_mean): Multiplication factor
**Problem**

`mean` corresponds to the expectation value of the annihilation
operators, and `xp_mean` corresponds to the expectation value of the
quadrature operators.

Using the equation
```math
a_j =  \frac{1}{\sqrt{2 \hbar}} (x_j + i p_j), \\
```
it can easily be realized, that the $` \sqrt{2 * \hbar} `$ should be
multiplied, not divided.

**Solution**

The `xp_mean` property got fixed, along with the corresponding tests.

References: #16