feat(GaussianState): Phase space rotation
The `rotated` method has been implemented to create a
phase-space-rotated version of the `GaussianState` instance.

The rotation in phase space is described by the equation

```math
x_{i, \phi} = \cos\phi~x_i + \sin\phi~p_i,
```

which corresponds to the following transformation in terms of the
annihilation operator

```math
a_i \mapsto e^{-i \phi} a_i.
```

Using this transformation, the `G` and `mean` attributes of
`GaussianState` are rotated, and a new instance is returned to avoid
altering the current state.

The `C` matrix is not needed to be updated, since the phase factors
cancel out in that case.

References: #16