feat(GaussianState): `reduced_rotated_mean_and_cov`
The `reduced_rotated_mean_and_cov` is implemented using the `reduced`
and `rotated` methods.

A transformed state is created by reducing the state to a subsystem
specified in `modes`, and then rotated in phase space by an angle `phi`.

The transformed state's quadratured ordered mean and covariance is
returned, as needed for the PennyLane plugin.

References: #16