feat(benchmarks): `vacuum_homodyne`
A simple benchmarking and profiling script has been added to the
newly-created `benchmarks` directory. This directory is excluded from
the package build.

For the simulations, 5 modes are used, and only vacuum states are
initialized, since squeezing is not (yet) implemented in `piquasso`. For
the gates, beamsplitters are used.

In `piquasso`, the `reduced_rotated_mean_and_cov` is calculated, which
is the same as `quad_expectation` in `strawberryfields`.

`timeit` is used for benchmarking, and `cProfile` for the profiling.
These are the simplest solutions.

`piquasso/__init__.py` has been extended with several imports needed for
the benchmark script, since it is desired to use this python package by
using it in the following fashion:
```
import piquasso as pq

state = pq.GaussianState(C=...,G=..., mean=...)
```
like the way `numpy` is used.

References: #39