feat(gaussian): Wigner function
The calculation of the Wigner function has been implemented under
`GaussianState`. Since the implementation got too large, the core of the
calcutation got moved to `functions.py` in the same folder.

The `gaussian_wigner_function` accepts not only a single canonical
coordinate vector, but a list of vectors.

**Note**

The `.flake8` file got extended with exclusions, because `flake8` ran on
those folders as well.