feat(operation): Creating an interface for modeless operations.
A class, `ModelessOperation` has been implemented.
It's subclasses are registered to `current_program.instructions` without having to call `__or__` of a `Q` mode object.

Like this, we have a nicer and less ambiguous syntax:

Instead of:
```
with self.program:
    Q(0, 1) | B(.5)
    Q() | Sampling(shots=10)
```
We have:
```
with self.program:
    Q(0, 1) | B(.5)
    Sampling(shots=10)
```

To be more elegant, `execute_instructions` has changed a little bit.

A test has been written for the registration.

Methods and tests has been updated according to changes.

Closes #42