refact(test_program): Removing odd dependencies
**Problem**

The current `TestProgram` test class is not `Backend` or
`Operation`-independent.

**Solution**

`TestProgram` got refactored with suitable, independent test dummy
fixtures.

To assert that the `Backend.execute_instructions` is called, a `Mock`
object has to be created, but `Program` expects a class instead of an
object. This is resolved with a small trick: a lambda function is passed
as `backend_class` to the `Program`.

The `blackbird`-related test `test_from_blackbird` is moved to
`TestBlackbirdParsing` for later resolution.