feat(registry): Storing classes by name
**Motivation**

In the future, we'll need to parse a `Program` instance from JSON. For
this, we need to be able to parse the classnames of the classes used in
piquasso. For example, we want to interpret `"GaussianBackend"` as the
`GaussianBackend` class.

**Solution**

`piquasso.registry` has been created. In it, a `ClassRecorder` class has
been defined, which needs to be subclassed in order to record the
subclasses which have been defined. This is done with
`__init_subclass__` as proposed in [PEP
487](https://www.python.org/dev/peps/pep-0487).

In the same file, a `retrieve_class` function has been created, which
retrieves the classes from the classnames.

A `State` has been created in a new `piquasso.state` module, which
subclasses `ClassRecorder`, and all our `*State` classes are subclassed
from this `State` base class. This way, all the state classes now can be
retrieved by plain strings.

Similarly, `Backend` and `Operation` got subclassed by `ClassRecorder`.

In `Program`, the possibility of specifying the `backend_class` via its
name has been added with the `Program._create_backend` method. This has
been tested, and the benchmark `vacuum_homodyne` has been rewritten to
use this.

References: #43