HOTFIX(backend): Remove mutation of iteration items
**Problem**

It broke `benchmarks/vacuum_homodyne.py`.

In `Backend.execute_instructions`, we iterated over a list, where we
altered the list items. This wouldn't be bad if the list items weren't
mutable, but they are. For this reason, the `instruction` name is
exactly the same as stored in the list. And this is bad, because Python
is not very happy about mutating list elements while looping, and can be
unpredictable.

**Solution**

The `.gitlab-ci.yml` has been extended with the `vacuum_homodyne`
benchmark.

In the `dict`s stored in instructions, all the keys that are not the
`"op"` key is moved under the `"kwargs"` key. The reason for this
naming is that the `dict` under `"kwargs"` are all the parameters which
are passed as keyword arguments to the method under `"op"`.

The related tests were refactored accordingly.