feat(Program): JSON parsing of `Program`s
**Motivation**

Somehow we need to create `Program` instances from JSON.

**Solution**

`Program.__init__` has been refactored to accept only classes and
instances (therefore `_create_backend` got deleted), and this logic is
moved to a new `from_properties` method, which can accept a mapping
to initialize a program.

Creating instances from mappings has been done using the `registry`,
where the `create_instance_from_mapping` function has been implemented,
which does exactly this using the classes registered to the `registry`.
For the initialization, a `from_properties` classmethod has to be
defined.

To handle `operations`, the `Program` is allowed to be initialized
with the set of operations in the form of a list of `Operation`
instances. This is also supported in the `from_properties` classmethod.

References: #43