refact(program): Instructions are operations
**Problems**

- There is no abstraction behind and `instruction`, and we handle these
  stuff everywhere around the code. It is hard to maintain consistency.

- `Operation`s are basically equivalent to the dicts we specify in several
  places and call `instruction`s, but without the `modes` parameter.

- In several places we refer to `op` or `operation` not as an
  `Operation` instance, but as a method of an `Operation` instance. This
  is weird and not trivial.

**Solution**

Let an `instruction` be equivalent to an `Operation` instance. We need
to let `Operation` instances know about `modes`. The biggest problem is
that the `Operation` instances are created by the user on the interface
with e.g. the following syntax:
```
Q(0, 1) | B(0.1, 0.3)
```

In order to overcome this, we need to specify `modes` later for the
instance. We can do it by using a `modes` attribute, which is set to
`None` by default.

To make `Operation` instances a bit more automatic, a `__call__` method
has been defined for them, which accepts a backend. This actually
resolves the method to be called corresponding to the `backend`
specified, and executes the operation. So in `execute_instructions` (now
called `execute_operations`, the `operation` instances just have to be
called in a for loop to execute the quantum program.

References: #43