fix(operations): `from_properties`
**Problem**

The `from_properties` initializes the `Operation` class with `params` as
the first argument, however, the `params` should be expaned with the `*`
operatior, as expected in `Operation.__init__`.

However, since often we initialize operations with (sometimes
compulsory) keyword arguments, the `**` expansion would be better. The
displacement operation from the UI wouldn't work.

**Solution**

That `**` operator is added in `from_properties`. A test case has been
added in `test_operations.py`, and `TestProgramJSONParsing` got fixed
(it was a false positive test) and extended with `params` as keyword
arguments.