HOTFIX(gaussian.state): Sign error
**Problem**

There were a sign error during the calculation of the covariance matrix.
This was noticed, when the square root of its determinant was calculated
for the Wigner function: the determinant was negative.

Even in the vacuum case, there were negative eigenvalues in the
covariance.

**Solution**

There was a sign error when the `corr` is calculated, which is the
second moment of the Gaussian state with respect to the quadratures.

Several tests got rewritten, and a new test has been written which
asserts for the covariance being proportional to the identity in the
case of a vacuum.