HOTFIX(gaussian.backend): Displacement gate
**Problem**

The displacement operation should transform the `C` and `G` matrices as
well. Previously, when the displacement operator got applied, the
determinant of the covariance matrix could turn negative, which is
non-physical: this value should always be non-negative.

**Solution**

The missing transformation of the `C` and `G` matrices got implemented.

Note, that neither `apply` nor `apply_active` could be used for this
case due to the nature of the displacement.