feat(gaussian): create_vacuum + Squeezing
A special `create_vacuum` method has been defined to ease the
instantiation of a vacuum Gaussian state. This has been included in the
benchmark.

The `sys.path` got appended in the benchmark script, because it resulted
in `ImportError`s under circumstances when the project is not installed
in the virtual environment.

The calculations got also extended with random squeezing and
displacement operations.

Finally, the `vacuum_homodyne` file got renamed to `homodyne`, because
it doesn't entirely operates on a vacuum now.