delete(package=fock): Fresh start for Fock module
The current `fock` package is quite useless, since it doesn't account
for multiple particles, has a very simple representation. It was just a
dummy implementation at the start of the development.

For this reason, this package has been deleted to enable a fresh start
for implementing the simulations in the Fock representation.

Issue: #54