refact(module=piquasso): Module wrapping
**Problem**

One might need to override certain classes (e.g. `GaussianState`) in the
implementation. This can not be achieved simply via the current
implementation.

**Solution**

The `piquasso` module is wrapped in a wrapper class called `Piquasso`.
Getting an attribute from the module `piquasso` from now on is done by
trying to retrieve items from the `registry`. If the retrieval is not
successful, then the original wrapped module is used for item retrieval.

In test files under `tests`, the `piquasso` module is used instead of
importing.

With this patch, the `piquasso` module would serve as the public-facing
interface of `piquasso`.