refact(operations): Backend method parametrization
**Problem**

There is a lot of similarity in the `*Backend` codes. For example, at
the beamsplitter operations, in both `SamplingBackend` and
`GaussianBackend` we calculate the 1-particle representation of the
gate.

**Solution**

We could avoid this redundancy by storing this logic in the
`Operation` classes, but first we need to pass the operation itself to
the `*Backend` methods.

Several tests has been refactored. Additionally, the
`piquasso/sampling/tests/functional/test_sampling.py` has been simply
moved to `tests/test_sampling.py`, because we usually store functional
tests there.

Some docstrings have been moved to `operations.py`, since the `*Backend`
classes are implementation details.