refact(operations): Unifying Backend implementations
The `*Backend.[phaseshift|beamsplitter]` methods got unified, because
they almost have the same functionality, the difference is in the
one-particle representation of the certain operations.

The `multiple_interferometer_on_modes` method got renamed to
`multiply_interferometer_on_modes`, because it makes more sense, and the
docstring also indicated that this method multiplies the interferometer
matrix.