refact(backend): Executing operations in backend
**Problem**

The way of resolving backend methods and executing the methods is
implemented in `Operation`, however, logically it is the job of the
`Backend`.

Also, if we want to specify additional flags regarding the execution, it
would be better to bring the implementation a bit closer to `Program`.

**Solution**

The operation executing logic is moved to `Backend` from `Operation`.