refact(operations): DI of operations and backends
**Problem**

Currently the `Operation` subclasses under `operations` depended on the
backends. This is somewhat odd, since the implemented operations on a
given `Backend` only concerns the `Backend` itself.

If one would have to specify the supported `Backend`s under the
`Operation` subclasses, it would be harder to e.g. create a new
`Backend`  from a plugin, since one would have to override the
`backends` attribute of every `Operation`.

**Solution**

The dependency got inverted: now the `Backend` subclasses depend on the
contents of `operations`, and not backwards. This has been done by
creating an abstract base class from `Backend` with an abstract method
called `get_operation_map`. The `_operation_map` would be used to get
the `Backend` method corresponding to an `Operation`, and would be
updated by `get_operation_map` per `*Backend` subclass.