refact(registry): Using plugins
**Problem**

There is no clear separation in the code, and the registration of
certian classes into the registry is somewhat hidden away.

On the other hand, making `piquasso` extensible would be great (see
`cpiquasso project).

Moreover, some classes which are not desired to be publicly exposed are
present in the registry as well, and the `Piquasso` plugin wrapper
exposes them too. This has to be avoided, a fine-grained control of the
registry is needed.

**Solution**

- The `__init_subclass__` method of `ClassRecorder` got deleted.

- `ClassRecorder` got renamed to `_Registry` (making it private), and
  `records` got renamed to `items` to avoid confusion.

- The `registry.add` method is added to use as a decorator at
  class definitions, e.g. at the `operations` module.

- The `registry.use_plugin` method is added to register plugins, which
  essentially registers all the classes from the `Plugin.classes`
  attribute to the registry.

- For plugins, a `Plugin` abstract base class has been created.

- The necessary contents of the `gaussian` and `sampling` packages has
  been collected into a `Plugin` instance called `DefaultPlugin`.

- The `from_properties` abstract part of the former `ClassRecorder` got
  refactoed into a mixin called `PropertyMixin`.