rename: Backend -> Circuit
**Problem**

The term `Backend` doesn't really provide much information about the
class.

`Backend` instances refer to classes which hold information about how to
apply the different `Operation` instances on the `State` instance. A
`Backend` instance also stores the `State` instance on which the
operation runs. The term `Circuit` is more suited for this.

**Solution**

All `Backend` and `backend` got renamed to `Circuit` and `circuit`,
respectively.