feat(pncfock): Particle number measurement
The particle number measurement has been implemented. This measurements
is stochastic, produces a random output, and also updates the state.
Being a projective measurement, the state is reduced to a pure state,
which is a Fock basis state in this case.

To provide a consistent way of presenting results, a `Result` class has
been created, which would be returned by `program.execute()`.