feat(program): Post-registration of state
Logically, the state creation in the program 'feels' better in the
program context with the operations: similar or connected things should
be collected together.

Therefore, the post-registration of a state is implemented.

All the test cases under `test_fock.py` has been rewritten with this
use-case, since it might be more relevant in Fock state preparation.