fix(constants): Custom confuguration of HBAR
**Problem**

In the previous implementation, the value of `hbar` could be set for the
program, but during execution, the executed code couldn't reach the
current program, because there isn't any. There is a `current_program`
in the context within the `with program` statement.

**Solution**

The value of `HBAR` module variable has been created to be set globally
by the user, and is used for the calculations.

The value `HBAR_DEFAULT` is set to private, not to be set by the user.