rename: PNCPure -> Pure
When using pure states, we actually represent the state the same way
regardless of using particle number conserving operators or not. Note,
that this may not be the case when representing the state with density
matrices, where certain matrix element can be omitted from the
calculation.

To reflect this, the `PNCPure*` classes and `pncpurefock` backend got
renamed to `Pure*` and `purefock`.

Some `test_backend.py` files got renamed to `test_circuit.py` in
compliance with commit `390fca4`.