feat(operations): Ladder operators
The creation and annihilation operators got implemented for the
`purefock` backend.

The `FockSpace` class got refactored to subclass a `tuple`. The basis
vectors of the Fock space are contained in this `tuple` object. This
step simplified much of the logic in `_math/fock.py`.

When some basis states overflow by having greater particle number than
the cutoff, those basis states are discarded. This would leave the state
with smaller a norm than before, therefore the state is normalized
again after using the ladder operators.