refact(test_fock): Using DMNumber
Previously, the `from_pure` method has been used to create a density
matrix. The usage of this is really inconvinient for the user, therefore
it got deleted, and all the tests were rewritten to use `DMNumber`
instead.

For the state preparation, a `__deepcopy__` method had to be added for
the `FockSpace` class, because `copy.deepcopy` couldn't copy a class
with both `__init__` and `__new__` for some reason.