feat(*fock): Particle number measurement on modes
The particle number (or Fock) measurement has been implemented for all
the `*fock` backends.

The tests have been extended by creating a new expected state according
to the measurement outcome. It might be questionable to test
stochastic events in the code, but monkeypatching `random.choices` would
mean testing the implementation, which is best avoided, too.

Note, that now the `from_number_preparations` is coupled with the
`Number` and `DMNumber` operations, this coupling might have to be
removed later in some way.

The implementation in `PNCFockState` and `PureFockState` has major
similarities in the procedure, only has differences in the underlying
representation. There might be a nice way to unify these, but in this
implementation, it is still not clear, and it might not even worth the
effort. Underlying implementation details of the `*State`'s have been
(mostly) put into `_math/fock.py`, however.