refact(tests): Storing test assets
**Problem**

There are a lot of (almost) meaningless test data e.g. for testing the
expected state of a simulation, making the tests unreadable and hard to
maintain: if some calculation changes, a whole lot of matrices has to be
rewritten.

**Solution**

Note, that it is not desired to create the the expected outputs with
`piquasso` itself, because that could hide errors. We need a fixed
expectation in the tests.

The expectations could be stored in an `assets` folder next to the test
file, and a tricky pytest fixture `assets` could be written to
handle the assets in a per-test basis. This returns with an instance of
`AssetHandler`, which could load and save asset files.

Since `AssetHandler` also has the capability of save the asset files
with the `save` method, it can be used to save assets when new test
cases are being written.

Note, that automatic resolution of the asset files might be convenient,
but prone to breaking at renaming and refactoring tests.

The `AssetHandler` is also used as a parent to
`GaussianStateAssetHandler`, which is used to load and save Gaussian
states by overriding the `load` and `save` methods.

**Note**

This might not be a good idea. This solution is really similar to the
`pytest-vcr` plugin, which can record requests/responses during testing
for storing them as assets. However, this system might be a bit of an
overkill.