feat(tests/gaussian): Storing mean and covariance
It might be better idea to store the mean and covariance in the tests,
because the underlying representation of the states might be irrelevant,
but the emergent mean and covariance properties are public, therefore
those should be tested.

This way, if the `GaussianState` undergoes a representation change, the
tests still shouldn't break.