refact(gaussian): Simplifying displacement
If we consider `C` and `G` to be elements in a covariance matrix instead
of a correlation matrix, the implementation for the displacement of the
Gaussian states could be majorly simplified.

This means, that the `cov` and `corr` properties on the `GaussianState`
would change dependency: the `cov` gets a bit more "fundamental" than
`corr`, `corr` would be computed from `cov`. The same line of reasoning
applies to `xp_cov` and `xp_corr`.