feat(gaussian): State initialization without m, G, C
The `GaussianState` class has been rewritten to avoid specifying the
`m`, `G` and `C` parameters at initialization since we need enforce the
user to prepare the state with regular operations. The underlying
representation is prone to future changes, and the user's code would
break easier.

Also, the `m`, `G` and `C` parameters have no direct use, it is quite
inconvenient to specify them.