feat(gaussian): MeasureParticleNumber
The particle number measurement is implemented in the `gaussian`
backend.

For this, a `hafnian` module is created under `_math`. This function
uses the power-trace method described in the first source.

In `GaussianState._apply_particle_number_measurement`, the `hafnian`
function is used to compute the probability distribution for generating
samples.

A benchmark has been added for the hafnian calculation and the particle
number measurement, too.

Note:

- The state is not altered, since it would not be Gaussian.

- The particle number measurement is slow, therefore the default cutoff
  is set to 3. In contrast, SF runs with a cutoff of 6. Also note, that
  the algorithm runs in exponential time.

Sources:

- [A faster hafnian formula for complex matrices and its benchmarking on
  a supercomputer](https://arxiv.org/pdf/1805.12498.pdf)
- [A detailed study of Gaussian Boson
  Sampling](https://arxiv.org/abs/1801.07488)