refact(api.operation): Keyword argument params
The `Operation.params` attribute is refactored to be a dict instead of a
tuple, since

- Handling of the `params` is non-trivial in the `Circuit` subclasses,
  since accessing the elements in the `params` tuple is done by integer
  indices.

- The ordering of the parameters coming from a JSON (see `apiquasso`)
  could be problematic. Changing of the ordering in `piquasso` would
  have consequences in `apiquasso` which tests might not catch.

A bunch of tests have been rewritten, too.

**BlackBird**

While refactoring the `Operation.params` attribute to a `dict`, the
BlackBird parsing has to be rewritten, too. This feature haven't even
taken account the possibility of strict keyword arguments, so e.g. the
`MZGate` couldn't be parsed.

The solution uses `inspect.signature`: using the signature of the
function, we could resolve the needed keyword arguments, and map the
BlackBird operation's (tuple) arguments to the keywords in order.

Note, that in CPython 3.6 the insertion ordereding of the `dict` was an
implementation detail, and fully becomes guaranteed in Python 3.7 as a
language feature. Since we support 3.6, `OrderedDict` is used, just to
make sure.