fix(Program): Blackbird loading with plugin
The Blackbird loading is moved under `core`, since it is quite
irrelevant from the other parts of the `Program` class.

Another issue is that the gates are loaded from
`piquasso.operations.gates`, however, if a plugin overrides the
a gate implementation, still the default gate is loaded. This has to be
rewritten to acquire classes from the `registry` itself.

Several test cases has been written. For convenience, the
`Operation.__eq__` and `Operation.on_modes` methods are defined, and the
`__repr__` method is fixed to work with `params` as a `dict`.

Moreover, the CX and CZ gater are imported in `piquasso/__init__.py` as
the other gates to avoid errors.

A warning regarding `collections.Iterable` at `api.circuit` has been
fixed.