fix(GBS): Husimi + Caching + Shortcut + Shots
**Husimi projection fix**

When calculating probabilites, the husimi matrix needs to be projected
for the given subspace defined by the modes, too.

**Caching**

Several calculations got refactored into an inline function for the
possibility of using `lru_cache` for these calculations. Most notable is
the `get_probability`, which caches the results of the hafnian
calculations for the case when multiple shots are used.

**Shortcut**

Originally, all the probabilities were calculated up to the cutoff, and
this probability distribution was used to generate a sample on a single
mode. This logic can be shortened by generating a number from a uniform
distribution on `(0, 1)` (called `guess`) and then stopping the
calculation when the cumulative probability reaches `guess`. If it is
not reached before the cutoff, we can use the already calculated
cumulative probability distribution to sample from that.

**Shots**

The shots are implemented by repeating the calculations by the value of
`shots`. The cache in `get_probability` is used, so that previous
calculations can be reused if needed.