feat(scripts): Distribution equality test
A hypothesis test has been created for multidimensional distribution
equality testing, which uses the `cramer` R library. No python package
found to be able to perform this test. Additionally, the `rpy2` package
should be installed to use R code from Python, but is not installed as a
development dependency, because there is no R in the pipeline (and
shouldn't be).

A histogram-creator script has also been added to visually compare the
distributions. For this, `matplotlib` has been added as a development
dependency.