fix(benchmarks): Benchmark setup fix
Some issues were found running the benchmarks. It turned out, that the
state of the SF `engine` and PQ `program` is persisted between runs,
yielding biased results.

In `gaussian_boson_sampling_benchmark` it wasn't a problem for PQ, since
the state is not altered at all, but SF encountered a strange
`KeyError`.

In `gaussian_boson_sampling_benchmark`, the issue was more severe,
therefore it had to be fixed.

The solution is to create and run the program in a single function, and
using the `benchmark` decorator to benchmark the whole process.

Two new fixtures were created for the mentioned benchmarks to return
only the example state, not the whole program (and engine).